Add: tools/isMobile
